+++ app/gui/src/project-view/components/ActionButton.vue
