+++ app/gui/src/project-view/components/ActionMenu.vue
