+++ app/gui/src/project-view/components/ContextMenu.vue
