+++ app/gui/src/project-view/components/ContextMenuTrigger.vue
