+++ app/gui/src/project-view/components/MenuButton.vue
@@ -62,7 +62,7 @@ function onClick() {
   }
   &.disabled {
     cursor: default;
-    opacity: 0.3;
+    opacity: 0.4;
     &:hover {
       background-color: unset;
     }
