+++ app/gui/src/project-view/components/MenuEntry.vue
